Split date picker from date picker input
