+++ catalog/date-picker-input/documentation.md
@@ -3,5 +3,5 @@ This documentation is automatically generated from jsdoc comments.
 ```react
 noSource: true
 ---
-<DocgenTable component={DatePicker} />
+<DocgenTable component={DatePickerInput} />
 ```
