+++ catalog/date-picker-input/variations.md
@@ -7,7 +7,7 @@ state: { selectedDate: null }
 <DatePickerDemo>
 <div>
 <span>The selected date is {dateFunctions.format(state.selectedDate, 'MM-dd-yyyy')}</span>
-<DatePicker
+<DatePickerInput
 	defaultSelectedDate={state.selectedDate || new Date()}
 	onChange={(date) => setState({ selectedDate: date })}
 	dateFunctions={dateFunctions}
