+++ components/date-picker/calendar-week/component.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { CalendarDate } from '../calendar-date/component.jsx';
-import { dateFunctionProps } from '../../date-function-props';
+import { dateFunctionProps } from '../date-function-props';
 
 const StyledCalendarWeek = styled.div`
 	display: flex;
