+++ components/date-picker/styled.jsx
@@ -1,43 +1,49 @@
 import styled from 'styled-components';
 import { colors } from '../shared-styles';
 
-export const Container = styled.div`
-	display: inherit;
-	flex-grow: 1;
-	position: relative;
-	max-width: 100%;
-	min-width: 0;
-	height: 32px;
-	font-size: 14px;
-	color: ${colors.flGray};
-`;
-
-export const Input = styled.input`
-	font-size: 1rem;
-	line-height: 1.3;
-	font-family: inherit;
-	color: @font-color;
-	border: 1px solid ${colors.borderColor};
-	border-radius: 3px;
-	height: 32px;
-	width: 100%;
+export const ChangeMonth = styled.button`
+	display: flex;
+	justify-content: space-around;
+	border: none;
 	margin: 0;
-	padding: 0.25rem 0.5rem;
-	transition: color 0.2s linear;
-	box-sizing: border-box;
-`;
-
-export const CalendarButton = styled.div`
-	position: absolute;
-	top: 0;
-	right: 7px;
+	padding: 0;
+	height: 34px;
+	width: 22px;
+	background: none;
 	cursor: pointer;
 `;
 
-export const CalendarIconContainer = styled.div`
-	padding: 7px 0;
+export const Header = styled.div`
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	white-space: nowrap;
+	border-bottom: none;
+	background: ${colors.white};
+	line-height: 32px;
+	font-weight: bold;
 `;
 
-export const DateTime = styled.div`
-	width: 204px;
+export const MonthLabel = styled.div`
+	display: inline-block;
+`;
+
+export const Week = styled.ul`
+	display: flex;
+	border-bottom: 1px solid ${colors.gray34};
+	padding: 8px 0;
+	background: ${colors.white};
+	list-style: none;
+	margin: 0;
+	color: ${colors.gray34};
+`;
+
+export const WeekDay = styled.li`
+	flex: 1;
+	text-align: center;
+	text-transform: uppercase;
+`;
+
+export const Month = styled.div`
+	background: ${colors.white};
 `;
