+++ components/main.js
@@ -16,4 +16,5 @@ export { Collapse } from './collapse/component.jsx';
 export { FilesSection } from './files-section/component.jsx';
 export { DropZone } from './drop-zone/component.jsx';
 export { Popover, PopoverManager, PopoverReference, Tooltip } from './popover';
+export { DatePickerInput } from './date-picker-input/component.jsx';
 export { DatePicker } from './date-picker/component.jsx';
